+++ src/components/views/symptom-checker/index.tsx
@@ -1,4 +1,4 @@
 export * from './consent';
 export * from './intro';
-export * from './quick-check-in';
+export * from './final';
 export * from './symptoms';
