+++ src/components/views/symptoms-history.tsx
@@ -1,27 +1,19 @@
-import React, {useState, useEffect, FC} from 'react';
-import {StyleSheet, View, Text, Linking} from 'react-native';
-import {
-  useFocusEffect,
-  useIsFocused,
-  useNavigation
-} from '@react-navigation/native';
+import React, {FC} from 'react';
+import {StyleSheet, View, Text} from 'react-native';
+import {useFocusEffect, useIsFocused} from '@react-navigation/native';
 import {useTranslation} from 'react-i18next';
 import {format} from 'date-fns';
 import {useExposure} from 'react-native-exposure-notification-service';
 
 import {useApplication} from 'providers/context';
-import {useSettings} from 'providers/settings';
 import {useAppState} from 'hooks/app-state';
-import {useSymptomChecker} from 'hooks/symptom-checker';
 
 import {Spacing, SingleRow} from 'components/atoms/layout';
 import {Card} from 'components/atoms/card';
 import {Heading} from 'components/atoms/heading';
-import {CheckInCard} from 'components/molecules/check-in-card';
-import {ResultCard} from 'components/molecules/result-card';
 import {Scrollable} from 'components/templates/scrollable';
 
-import {Symptom, SymptomRecord, SymptomsCheckResult} from 'constants/symptoms';
+import {Symptom, SymptomRecord} from 'constants/symptoms';
 import {text, colors} from 'theme';
 import {BubbleIcons} from 'assets/icons';
 
@@ -30,29 +22,12 @@ interface SymptomListItem {
   label: string;
 }
 
-function countSymptoms(symptoms: SymptomRecord): number {
-  return Object.values(symptoms).reduce((t: number, r: number) => t + r, 0);
-}
-
-export const SymptomsHistory: FC<any> = ({route}) => {
-  const [
-    checkInResult,
-    setCheckInResult
-  ] = useState<SymptomsCheckResult | null>(null);
-
+export const SymptomsHistory: FC<any> = () => {
   const {t} = useTranslation();
-  const navigation = useNavigation();
   const isFocused = useIsFocused();
   const [appState] = useAppState();
-  const {getNextScreen} = useSymptomChecker();
 
-  const {ageRangeOptions, checkerThankYouText} = useSettings();
-  const {
-    user,
-    completedChecker,
-    checks,
-    verifyCheckerStatus
-  } = useApplication();
+  const {checks, verifyCheckerStatus} = useApplication();
   const {readPermissions} = useExposure();
 
   useFocusEffect(
@@ -66,67 +41,8 @@ export const SymptomsHistory: FC<any> = ({route}) => {
     }, [isFocused, appState, verifyCheckerStatus])
   );
 
-  useEffect(() => {
-    const {symptoms, feelingWell} = route.params?.feelingWell
-      ? route.params
-      : checks[0];
-    let result: SymptomsCheckResult;
-
-    const symptomsCount = countSymptoms(symptoms);
-    if (!symptomsCount) {
-      if (
-        ageRangeOptions.find((g) => g.value === user?.ageRange && g.riskGroup)
-      ) {
-        result = 'riskGroup';
-      } else {
-        result =
-          checks[0] && countSymptoms(checks[0].symptoms || {}) > 0
-            ? 'recovered'
-            : feelingWell
-            ? 'noSymptomsFeelingWell'
-            : 'noSymptomsNotFeelingWell';
-      }
-    } else {
-      result = 'coronavirus';
-    }
-    setCheckInResult(result);
-  }, []);
-
   return (
     <Scrollable safeArea={false} backgroundColor={colors.white}>
-      {completedChecker && checkInResult && (
-        <>
-          <Heading
-            accessibilityRefocus
-            accessibilityFocus
-            text={t('checker:results:title')}
-          />
-          <ResultCard
-            message={checkerThankYouText[checkInResult]}
-            markdownProps={{
-              style: {},
-              warningList: true
-            }}
-            buttonText={t('checker:results:stopTheSpread')}
-            onButtonPress={() =>
-              Linking.openURL(
-                'https://www.health.pa.gov/topics/disease/coronavirus/Pages/Stop-the-Spread.aspx'
-              )
-            }
-          />
-        </>
-      )}
-      {!completedChecker && (
-        <CheckInCard
-          accessibilityRefocus
-          accessibilityFocus
-          onPress={() =>
-            navigation.navigate('symptoms', {
-              screen: getNextScreen()
-            })
-          }
-        />
-      )}
       <Spacing s={20} />
       <Heading
         accessibilityRefocus
